Seperated the kisok app install broadcast receiver
